requirements.txt
numpy==2.2.6
pygame==2.6.1
PyOpenGl==3.1.9
pillow==11.2.1